Hide iris.analysis.interpolate behind deprecation wrapper; tests show equivalence to newer methods.
